ENH: add multiway-clustering for PanelOLS
add a function (multi_way_cluster_iter) to calculate the sum(and subtractions) of the clustered covariance matrices to be added or subtracted. Added when an odd number of cluster dimensions, and subtracted when even.

add a correction (cgm_vcov_fix) when non-positive semi-definite VCOV matrix

call multi_way_cluster_iter from the cov method of panel.covariance.ClusteredCovariance